v0.2.302

+++ server/graphql/types/transport-order-detail/transport-order-detail.ts
@@ -12,7 +12,7 @@ export const TransportOrderDetail = gql`
     deliveryOrder: DeliveryOrder
     transportDriver: TransportDriver
     transportVehicle: TransportVehicle
-    assignedLoad: Int
+    assignedLoad: Float
     updater: User
     creator: User
     updatedAt: String
